Restructured barn kit directory
