+++ farmdata2/farmdata2_modules/fd2_barn_kit/transplantingReport/transplantingReport.spec.js
@@ -1,12 +1,11 @@
 const dayjs = require('dayjs')
-var FarmOSAPI = require('../resources/FarmOSAPI.js')
+var FarmOSAPI = require('../../resources/FarmOSAPI.js')
 var getSessionToken = FarmOSAPI.getSessionToken
 var getCropToIDMap = FarmOSAPI.getCropToIDMap
 var getIDToCropMap = FarmOSAPI.getIDToCropMap
 var getAreaToIDMap = FarmOSAPI.getAreaToIDMap
 var getUserToIDMap = FarmOSAPI.getUserToIDMap
 var getUnitToIDMap = FarmOSAPI.getUnitToIDMap
-var createRecord = FarmOSAPI.createRecord
 var getRecord = FarmOSAPI.getRecord
 var deleteRecord = FarmOSAPI.deleteRecord
 var getConfiguration = FarmOSAPI.getConfiguration
